not overwriting existing dirs

+++ setup_all.py
@@ -3,15 +3,31 @@ import os,sys,glob,subprocess, urllib2, tarfile, shutil
 
 failures = []
 
-def install_url(f, url, subdir=None, plugin=True):
+
+def get_options():
+    import os, os.path
+    from optparse import OptionParser
+    parser = OptionParser()    
+    parser.add_option("-f", "--force", dest="force", help="reinstall even existing plugins", action="store_true", default=False)
+
+    (options, args) = parser.parse_args()
+
+    return options, args
+
+def install_url(f, url, subdir=None, plugin=True, force=False):
     print "install url: package=", f, "url= ", url
+
+    if os.path.exists(f):
+        if force:
+            shutil.rmtree(f)
+        else:
+            print "Path exists: ", f
+            return;
+
     response = urllib2.urlopen(url)
     thetarfile = tarfile.open(fileobj=response, mode="r|gz")
     thetarfile.extractall()
-
-    if os.path.exists(f):
-        shutil.rmtree(f)
-
+    
     dirname = glob.glob("*%s*" % (f))
     curdir = os.getcwd()
     dirname = dirname[0]
@@ -38,6 +54,8 @@ def install_url(f, url, subdir=None, plugin=True):
     os.chdir(curdir)
 
 
+options, args = get_options()
+
 #files = ["Turbine_CostsSE", "CommonSE", "Plant_CostsSE", "Plant_FinanceSE", "Plant_EnergySE"]
 files = ["Turbine_CostsSE", "CommonSE", "Plant_CostsSE", "Plant_FinanceSE", "Plant_EnergySE",
          "AeroelasticSE", "AirfoilPreppy", "CCBlade", "DriveSE", "DriveWPACT", "NREL_CSM", "RotorSE",
@@ -63,16 +81,16 @@ f = "fusedwind"
 #subdir = os.path.join(f,f) # fusedwind is nested!... not anymore, I guess
 subdir = f
 url = "http://github.com/FUSED-Wind/fusedwind/tarball/develop"  ## note, develop branch
-install_url(f,url)
+install_url(f,url,force=options.force)
 
 #install Andrew Ning's akima
 url = "http://github.com/andrewning/akima/tarball/master"
-install_url("akima", url, plugin=False)
+install_url("akima", url, plugin=False,force=options.force)
 
 # download and install all the necessary WISDEM plugins
 for f in files:
     url = "%s%s/tarball/master" % (wis, f)
-    install_url(f, url)
+    install_url(f, url,force=options.force)
     
 # finally install WISDEM itself
 os.chdir(rootdir)
